fix(mindtagger): show the grouping filter when `grouping_columns` is defined
previously, if the `grouping_columns` config is defined, but no `groupFilter` is set
(i.e. the `g` parameter on the URL is empty), then the mindtagger UI would not show
the grouping filter element, which then prevents the `groupFilter` from being set.

this patch shows the grouping filter by default if `MindtaggerTask.config.items.grouping_columns`
is defined, allowing the users to freely set their group filter from the UI.